backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1

sse-starlette==1.8.2
asyncio-mqtt==0.16.1
